Refactor request parsing
